Changed variable width decode method
Changed the compaction of bins to decode at 20 Hz resolution.  Seems
better now.  Still watching to get a better feel for it but for the
returns I'm seeing it's worth losing a decode or two that other might
have gotten.  Long duration it might miss 1% of what other would - and
90% of the phantom/alias/garbage decodes.